requirements-simple.txt
pandas==2.2.2
pytrends==4.9.2
psycopg2-binary==2.9.9
scipy==1.14.1
numpy==2.1.3
